やり取りで教わったところ修正

+++ src/core/app/mentorapp/create_mentor_requests.go
@@ -5,13 +5,13 @@ import "time"
 type CreateMentorRecruitmentRequest struct {
 	Title                 string    `json:"title"`
 	CategoryName          string    `json:"category_name"`
-	BudgetFrom            int       `json:"budget_from"`
-	BudgetTo              int       `json:"budget_to"`
+	BudgetFrom            uint32    `json:"budget_from"`
+	BudgetTo              uint32    `json:"budget_to"`
 	ApplicationPeriodFrom time.Time `json:"application_period_from"`
 	ApplicationPeriodTo   time.Time `json:"application_period_to"`
-	ConsultationFormat    int       `json:"consultation_format"`
-	ConsultationMethod    int       `json:"consultation_method"`
+	ConsultationFormat    uint8     `json:"consultation_format"`
+	ConsultationMethod    uint8     `json:"consultation_method"`
 	Description           string    `json:"description"`
-	Status                int       `json:"status"`
+	Status                uint8     `json:"status"`
 	TagNames              []string  `json:"tag_names"`
 }
